add shooting for character

+++ Assets/TaskFrom1C/Enemy/Wave/WaveConfig.cs
@@ -14,7 +14,7 @@ namespace TaskFrom1C.Enemy
     public struct EnemyData
     {
         public float Speed;
-        public float Health;
-        public float Damage;
+        public int Health;
+        public int Damage;
     }
 }
